requirements.txt
streamlit>=1.32.0
psycopg2-binary>=2.9.9
pandas>=2.2.0
python-dotenv>=1.0.0
thefuzz>=0.19.0
python-Levenshtein>=0.21.1
Faker>=22.6.0
matplotlib>=3.8.0
